src/user_stories/job_seeker.py
```python
from src.automated_followup import set_follow_up
from src.personalization_engine import generate_personalized_content

class JobSeeker:
    def __init__(self, user_email, user_password):
        self.user_email = user_email
        self.user_password = user_password
        self.trigger_conditions = {"days_after": 7}
        self.follow_up_content = None

    def set_follow_up_trigger(self):
        set_follow_up(self.user_email, self.user_password, self.trigger_conditions)

    def generate_follow_up_content(self):
        self.follow_up_content = generate_personalized_content(self.user_email, "job_application_follow_up")
        return self.follow_up_content

    def send_follow_up_email(self):
        if self.follow_up_content is None:
            self.generate_follow_up_content()
        send_follow_up(self.user_email, self.user_password, self.follow_up_content)
```